+++ src/components/noteContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const NoteContext = createContext();
+
+export default NoteContext;
